Add *ready* plugin
Add a ready plugin that allows plugin to signal when they are ready.
This uses simple functions defined in the ready plugin; RegisterPlugin
and Signal.

Add README.md and tests moduled on the health plugin; which will be
relegated to just providing process health. In similar vein to health
this is a process wide setting meaning it will need to include all
plugins for all server block when signalling readiness.

During startup it block forever, while displaying the status every
second:

~~~
2019-02-25T16:09:02.168Z [INFO] plugin/ready: Waiting on 2/2: erratic
2019-02-25T16:09:03.168Z [INFO] plugin/ready: All plugins signal readiness
.:1053
bla.:1053
2019-02-25T16:09:03.168Z [INFO] CoreDNS-1.3.1
2019-02-25T16:09:03.168Z [INFO] linux/amd64, go1.11.4,
CoreDNS-1.3.1
linux/amd64, go1.11.4,
^C[INFO] SIGINT: Shutting down
~~~

This was started with this Corefile
~~~
. {
    erratic
    whoami
    ready
}

bla {
    erratic
    whoami
}
~~~

Both *erratic* plugins are reporting readiness.

Signed-off-by: Miek Gieben <miek@miek.nl>

map can be [string]int

Signed-off-by: Miek Gieben <miek@miek.nl>

shorter

Signed-off-by: Miek Gieben <miek@miek.nl>